Created a base entity class to prevent duplications with the equal and hashcode methods
